HSM-14-19 decomposition cont.: carve ReviewUI (app 2,880 -> 1,797, -78% from the start)
ReviewUI.swift (1,083) — the on-device artifact generation + REVIEW surface (HSM-8-04):
MeetingReviewState (the generation theater + MIR routing), the swipeable artifact cards + detail +
voice-correction sheet, the transcript view, segment replay, and the shared artifact/profile
formatting helpers. Needed `import PencilKit` (PKDrawing) + GenerationTheater widened to internal
(used by MeetingDetailView) — both caught by the verified build.

Device-arch xcodebuild BUILD SUCCEEDED. Ten modules carved; MeetingCaptureApp.swift 8,042 -> 1,797.
The monster is now the MeetingApp core (entry + CaptureModel + the main screens) — the clean base
the real DeskOS gets built on. Decomposition was the plumbing; the mission is the experience.
